More discreet sharing of db-conn among system test runner threads.
When running a transactional system test, the web service thread
needs to use the same DB connection as the thread that sets up the
fixtures.  This code adds a "using_connection" method to ConnectionPool
which allows one thread to "borrow" the connection of another, without
affecting other threads that just want to be left alone.